move back to original path
